[EPAD8-2249] Refactoring actions to utilize a shared base.
